+++ ChargeBee/Models/Card.cs
@@ -88,6 +88,10 @@ namespace ChargeBee.Models
         {
             get { return GetValue<string>("billing_city", false); }
         }
+        public string BillingStateCode 
+        {
+            get { return GetValue<string>("billing_state_code", false); }
+        }
         public string BillingState 
         {
             get { return GetValue<string>("billing_state", false); }
@@ -168,6 +172,11 @@ namespace ChargeBee.Models
             public UpdateCardForCustomerRequest BillingCity(string billingCity) 
             {
                 m_params.AddOpt("billing_city", billingCity);
+                return this;
+            }
+            public UpdateCardForCustomerRequest BillingStateCode(string billingStateCode) 
+            {
+                m_params.AddOpt("billing_state_code", billingStateCode);
                 return this;
             }
             public UpdateCardForCustomerRequest BillingState(string billingState)